Updated client.c to scan and connect to devices I made new_client.c be able to scan for decices once the client program starts
This will be very helpful in the future so that a single client will be
able to connect to multiple malicious bluetooth servers to conduct
attacks

This way we can have multiple raspberry pi's around an area that are out
of the reach of bluetooth so we can expand our area of attack!

I all add changes to the shell script verysoon!